Move forking logic to Dependency Locator and wrapper classes
